# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi primera documentacion con swagger",
        "description": "Si",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Lista todas los categorias.",
                "operationId": "697131dd052703b53059fbe02b9decbd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_categoria": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "electronicos"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Todo tipo de electronicos"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Crea una nueva categoria",
                "description": "Crea una nueva categoria con los datos proporcionados en el cuerpo de la solicitud.",
                "operationId": "9559d9d92bcad5dc09d658e921badd50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Descripcion"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Ropa"
                                    },
                                    "Descripcion": {
                                        "type": "string",
                                        "example": "Ropa y accesorios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obtiene una categoria por su ID",
                "description": "Devuelve una categoria específico según el ID proporcionado.",
                "operationId": "a090c50b0f725bf25d75c01aa136cb6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Categoria": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Actualiza una nueva categoria",
                "description": "Actualiza una nueva categoria con los datos proporcionados en el cuerpo de la solicitud.",
                "operationId": "9aa5ce8b93543a83fd190257a3aba297",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Descripcion"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Ropa"
                                    },
                                    "Descripcion": {
                                        "type": "string",
                                        "example": "Ropa y accesorios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Borra una  categoria",
                "description": "Borra una  categoria con los datos proporcionados en el cuerpo de la solicitud.",
                "operationId": "eeac9b2988ff4c638cfd3e83f602a468",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categoria": {
                "title": "Categoria",
                "description": "Modelo de tabla Categoria",
                "properties": {
                    "id_categoria": {
                        "type": "integer",
                        "example": "5"
                    },
                    "Nombre": {
                        "type": "string",
                        "example": "Marca Gamma alta"
                    },
                    "Descripcion": {
                        "type": "string",
                        "example": "La marca mascara de todas."
                    }
                },
                "type": "object"
            },
...
            "Pedidos": {
                "title": "Pedidos",
                "description": "Modelo de datos para Pedidos.",
                "properties": {
                    "id_pedido": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fecha_creacion": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-02-23T00:09:16.000000Z"
                    },
                    "id_proveedor": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_forma_pago": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_cliente": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_estado_pedido": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    }
}